🧹 [refactor async_pool into package]
🎯 What: Refactored the monolithic 858-line `async_pool.py` file into a well-structured `async_pool` package.
💡 Why: The original file contained multiple responsibilities (models, executors, processors, connection pooling), violating the Single Responsibility Principle and making it hard to maintain. Extracting these into separate cohesive modules improves maintainability and readability.
✅ Verification: Confirmed that all `pytest tests/` in `packages/workhorse` pass and no functionality was broken. Preserved backward compatibility by correctly mapping the original exported properties in `async_pool/__init__.py`.
✨ Result: `async_pool` logic is now neatly categorized into `models`, `connection_pool`, `executors`, `processors`, and `utils` sub-modules.
