+++ packages/workhorse/src/settler_workhorse/async_pool/__init__.py
@@ -1,0 +1,25 @@
+"""Async connection pooling and concurrent job processing.
+
+Provides database connection pooling, async job processing, and
+backpressure handling for high-throughput ML/AI workloads.
+"""
+
+from .connection_pool import ConnectionInterface, ConnectionPool
+from .executors import AsyncTaskExecutor
+from .models import PoolState, PoolStats, TaskStats
+from .processors import AsyncJobProcessor, StreamingProcessor
+from .utils import batch_process, parallel_map, sync_to_async
+
+__all__ = [
+    "PoolState",
+    "PoolStats",
+    "TaskStats",
+    "ConnectionInterface",
+    "ConnectionPool",
+    "AsyncTaskExecutor",
+    "AsyncJobProcessor",
+    "StreamingProcessor",
+    "parallel_map",
+    "batch_process",
+    "sync_to_async",
+]
